+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -2,6 +2,7 @@ package com.pinterest.ktlint.ruleset.experimental
 
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
+import com.pinterest.ktlint.ruleset.experimental.trailingcomma.TrailingCommaRule
 
 public class ExperimentalRuleSetProvider : RuleSetProvider {
 
@@ -11,6 +12,7 @@ public class ExperimentalRuleSetProvider : RuleSetProvider {
         ArgumentListWrappingRule(),
         MultiLineIfElseRule(),
         NoEmptyFirstLineInMethodBlockRule(),
+        TrailingCommaRule(),
         PackageNameRule(),
         EnumEntryNameCaseRule(),
         SpacingAroundDoubleColonRule(),